requirements.txt
asgiref==3.5.0
cffi==1.15.0
cryptography==36.0.1
Django==4.0.3
django-rest-framework==0.1.0
django-rest-knox==4.2.0
djangorestframework==3.13.1
djongo==1.3.6
mysqlclient==2.1.0
pycparser==2.21
pymongo==3.12.3
pytz==2021.3
sqlparse==0.2.4
tzdata==2021.5
